refactor: 아이템을 수정하는 기능 분리
